调整线路详情-站点详情

+++ lib/entity/station_real_time_info_entity.dart
@@ -98,7 +98,7 @@ class RealtimeInfoList {
   @JsonKey(name: 'DepartureState')
   final String? departureState;
   @JsonKey(name: 'SortTime')
-  final int? sortTime;
+  final double? sortTime;
 
   const RealtimeInfoList({
     this.busID,
